backend/requirements.txt
Flask==3.0.2
supabase==2.4.3
PyYAML==6.0.1
requests==2.31.0
Flask-Cors==4.0.0
python-dotenv==1.0.1